pyproject.toml
[project]
name = "landing-page"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django==5.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
]
prod = [
    "gunicorn>=23.0.0",
    "whitenoise>=6.11.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"**/settings/*.py" = ["F403", "F405"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
